+++ CHANGELOG.rst
@@ -17,9 +17,21 @@ Unreleased
 ----------
 None
 
-[3.53.1]
+[3.53.4]
 --------
 fix: update course run selection logic for SAP content exporter
+
+[3.53.3]
+--------
+feat: mark Cornerstone session token fields read-only in admin
+
+[3.53.2]
+--------
+feat: update data sharing consent request language
+
+[3.53.1]
+--------
+feat: Django Admin tweaks for integrations
 
 [3.53.0]
 --------